Moved the ServerInfo class from codeu.chat.util to codeu.chat.common
