+++ src/codeu/chat/server/Server.java
@@ -29,6 +29,7 @@ import codeu.chat.common.NetworkCode;
 import codeu.chat.common.Relay;
 import codeu.chat.common.Secret;
 import codeu.chat.common.User;
+import codeu.chat.common.ServerInfo;
 import codeu.chat.util.*;
 import codeu.chat.util.connections.Connection;
