+++ src/codeu/chat/server/View.java
@@ -24,8 +24,8 @@ import codeu.chat.common.ConversationPayload;
 import codeu.chat.common.Message;
 import codeu.chat.common.SinglesView;
 import codeu.chat.common.User;
+import codeu.chat.common.ServerInfo;
 import codeu.chat.util.Logger;
-import codeu.chat.util.ServerInfo;
 import codeu.chat.util.Uuid;
 import codeu.chat.util.store.StoreAccessor;
